pyproject.toml
[project]
name = "de-airflow-pipeline"
version = "0.1.0"
description = "Salesforce to PostgreSQL data pipeline with Airflow"
requires-python = ">=3.11"
dependencies = [
    "apache-airflow[kubernetes]==2.8.1",
    "apache-airflow-providers-postgres==5.10.0",
    "apache-airflow-providers-salesforce==5.6.0",
    "pandas==2.1.4",
    "psycopg2-binary==2.9.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["de_airflow_pipeline"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "S", "A", "C4", "T20"]
ignore = ["S101"]  # Allow assert statements in tests

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
]

[tool.coverage.run]
source = ["de_airflow_pipeline"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
]